Fixed bugs in order classes and orderservice, changed orderline slightly, made test for order, orderservice

+++ src/main/java/avansivh11/dehartigesupermarkt/model/product/Product.java
@@ -31,9 +31,9 @@ public class Product {
 
     @NotNull(message = "Price is required.")
     @Min(0)
-    private int price;
+    private double price;
 
-    public Product(String name, String description, int price) {
+    public Product(String name, String description, double price) {
         this.name = name;
         this.description = description;
         this.price = price;
